Enable VM integration tests by building from unified sibling sources
The NixOS VM integration tests were disabled (gated behind `&& false`)
because they require building aspen-node and aspen-cli, which depend on
~38 extracted sibling repos not available in the Nix sandbox.

This adds a full-source build pipeline that assembles the main workspace
with all sibling repos into a unified source tree (.nix-deps/), rewrites
all relative paths, inlines workspace = true references, converts git deps
to path deps, and merges Cargo.lock files. The builds require `--impure`
since sibling repos live outside the flake source tree.

The pipeline includes:
- inline-workspace-deps.sh: resolves workspace = true references from
  each crate's actual workspace root, handling multi-line values and
  feature merging
- git-to-path-deps.py: converts GitHub git deps to local path deps
  pointing at the unified source tree
- merge-cargo-locks.py: merges [[package]] entries from sibling locks
- full-build-Cargo.lock: pre-generated unified lock file covering all
  sibling repo transitive deps

Also fixes test failures from the prior test suite pass:
- DeterministicKeyValueStore: implement Batch/ConditionalBatch/Transaction
  (was silently dropping writes via catch-all)
- Directory layer: handle Ok(ReadResult { kv: None }) as key-not-found
  instead of treating as empty value
- Feature gating: wire aspen-testing-madsim/sql and jobs features
- Skip raft verus specs (extracted to aspen-raft repo)

+++ crates/aspen-testing/src/job_worker_tester.rs
@@ -447,7 +447,7 @@ impl JobWorkerTester {
             }
 
             // Start workers
-            pool.start(config.workers_per_node).await.expect("failed to start workers");
+            pool.start(config.workers_per_node as u32).await.expect("failed to start workers");
 
             job_managers.push(manager);
             worker_pools.push(pool);
